+++ code/src/Plexdata.LogWriter.Abstraction.Tests/Internals/Factories/LoggerFactoryTests.cs
@@ -1,7 +1,7 @@
 ﻿/*
  * MIT License
  * 
- * Copyright (c) 2019 plexdata.de
+ * Copyright (c) 2021 plexdata.de
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -77,7 +77,6 @@ namespace Plexdata.LogWriter.Abstraction.Tests.Internals.Factories
             Assert.That(this.instance.CreateLogEventFormatter(this.settings), Is.InstanceOf(typeof(RawFormatter)));
         }
 
-        [Test]
         [TestCase(LogType.Raw, typeof(RawFormatter))]
         [TestCase(LogType.Csv, typeof(CsvFormatter))]
         [TestCase(LogType.Json, typeof(JsonFormatter))]