+++ code/src/Plexdata.LogWriter.Abstraction.Tests/Logging/EmptyLoggerDefaultTests.cs
@@ -1,7 +1,7 @@
 ﻿/*
 * MIT License
 * 
-* Copyright (c) 2019 plexdata.de
+* Copyright (c) 2021 plexdata.de
 * 
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
@@ -46,7 +46,6 @@ namespace Plexdata.LogWriter.Abstraction.Tests.Logging
 
         #region IsEnabled
 
-        [Test]
         [TestCase(LogLevel.Disabled)]
         [TestCase(LogLevel.Trace)]
         [TestCase(LogLevel.Debug)]