+++ code/src/Plexdata.LogWriter.Abstraction/Internals/Extensions/TypeFormatterExtension.cs
@@ -1,7 +1,7 @@
 ﻿/*
  * MIT License
  * 
- * Copyright (c) 2019 plexdata.de
+ * Copyright (c) 2021 plexdata.de
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -136,7 +136,7 @@ namespace Plexdata.LogWriter.Internals.Extensions
                 type = value?.GetType();
             }
 
-            return type == null ? false : TypeFormatterExtension.mappings.ContainsKey(type);
+            return type != null && TypeFormatterExtension.mappings.ContainsKey(type);
         }
 
         /// <summary>