+++ code/src/Plexdata.LogWriter.Console.Tests/Definitions/DimensionTests.cs
@@ -1,7 +1,7 @@
 ﻿/*
  * MIT License
  * 
- * Copyright (c) 2019 plexdata.de
+ * Copyright (c) 2021 plexdata.de
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -160,7 +160,6 @@ namespace Plexdata.LogWriter.Console.Tests.Definitions
             Assert.That(dimension.GetHashCode(), Is.EqualTo(hash));
         }
 
-        [Test]
         [TestCase(-1, -1, "IsValid=False, Width=-1, Lines=-1")]
         [TestCase(-1, 0, "IsValid=False, Width=-1, Lines=0")]
         [TestCase(0, -1, "IsValid=False, Width=0, Lines=-1")]