+++ code/src/Plexdata.LogWriter.Mail.Tests/Settings/MailLoggerSettingsTests.cs
@@ -1,7 +1,7 @@
 ﻿/*
  * MIT License
  * 
- * Copyright (c) 2019 plexdata.de
+ * Copyright (c) 2021 plexdata.de
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -121,7 +121,6 @@ namespace Plexdata.LogWriter.Mail.Tests.Settings
             Assert.That(instance.LogType, Is.EqualTo(LogType.Raw));
         }
 
-        [Test]
         [TestCaseSource(nameof(EncodingTestItemList))]
         public void MailLoggerSettings_ConfigurationValid_GetSectionValueForEncodingAsExpected(Object current)
         {