+++ code/src/Plexdata.LogWriter.Stream.Tests/Settings/StreamLoggerSettingsTests.cs
@@ -1,7 +1,7 @@
 ﻿/*
  * MIT License
  * 
- * Copyright (c) 2019 plexdata.de
+ * Copyright (c) 2021 plexdata.de
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -132,7 +132,6 @@ namespace Plexdata.LogWriter.Stream.Tests.Settings
             Assert.That(instance.LogType, Is.EqualTo(LogType.Json));
         }
 
-        [Test]
         [TestCaseSource(nameof(EncodingTestItemList))]
         public void StreamLoggerSettings_ConfigurationValid_GetSectionValueForEncodingAsExpected(Object current)
         {